fix - 마이페이지 및 로그인 응답 정보 추가
