+++ server/src/main/java/challenge/server/config/SecurityConfig.java
@@ -11,8 +11,6 @@ import challenge.server.security.handler.UserAuthenticationFailureHandler;
 import challenge.server.security.handler.UserAuthenticationSuccessHandler;
 import challenge.server.security.jwt.JwtTokenizer;
 import challenge.server.security.utils.CustomAuthorityUtils;
-import challenge.server.user.repository.LogoutListRepository;
-import challenge.server.user.repository.UserRepository;
 import challenge.server.user.service.LogoutListService;
 import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -50,7 +48,6 @@ public class SecurityConfig { // https 적용
     private final CustomOAuth2UserService customOAuth2UserService;
     private final UserService userService;
     private final LogoutListService logoutListService;
-    private final UserRepository userRepository;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -67,7 +64,7 @@ public class SecurityConfig { // https 적용
                 .authenticationEntryPoint(new UserAuthenticationEntryPoint()) //Oauth2에서는 인증에서 실패했을때 처리하는 로직
                 .accessDeniedHandler(new UserAccessDeniedHandler()) //인가 에러 핸들링
                 .and()
-                .apply(new CustomFilterConfigurer(jwtTokenizer, authorityUtils, userService, logoutListService, userRepository))
+                .apply(new CustomFilterConfigurer())
                 .and()
                 // v1
                 .authorizeHttpRequests(authorize -> authorize.anyRequest().permitAll())
@@ -156,6 +153,25 @@ public class SecurityConfig { // https 적용
         UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
         source.registerCorsConfiguration("/**", configuration);
         return source;
+    }
+
+    public class CustomFilterConfigurer extends AbstractHttpConfigurer<CustomFilterConfigurer, HttpSecurity> {
+        @Override
+        public void configure(HttpSecurity builder) throws Exception {
+            AuthenticationManager authenticationManager = builder.getSharedObject(AuthenticationManager.class);
+//            builder.addFilter(new CustomFilter(authenticationManager));
+
+            JwtAuthenticationFilter jwtAuthenticationFilter = new JwtAuthenticationFilter(authenticationManager, jwtTokenizer, userService);
+            jwtAuthenticationFilter.setFilterProcessesUrl("/login");
+            jwtAuthenticationFilter.setAuthenticationSuccessHandler(new UserAuthenticationSuccessHandler());
+            jwtAuthenticationFilter.setAuthenticationFailureHandler(new UserAuthenticationFailureHandler());
+
+            JwtVerificationFilter jwtVerificationFilter = new JwtVerificationFilter(jwtTokenizer, authorityUtils, logoutListService);
+
+            builder
+                    .addFilter(jwtAuthenticationFilter)
+                    .addFilterAfter(jwtVerificationFilter, JwtAuthenticationFilter.class);
+        }
     }
 
 //    @Bean
