+++ server/src/main/java/challenge/server/security/filter/JwtVerificationFilter.java
@@ -39,38 +39,17 @@ import java.util.stream.Collectors;
 
 @Component
 @Slf4j
+@RequiredArgsConstructor
 public class JwtVerificationFilter extends OncePerRequestFilter {
     private final JwtTokenizer jwtTokenizer;
     private final CustomAuthorityUtils authorityUtils;
     private final LogoutListService logoutListService;
-    private final AuthenticationManager authenticationManager;
 
-    public JwtVerificationFilter(JwtTokenizer jwtTokenizer, CustomAuthorityUtils authorityUtils, LogoutListService logoutListService, AuthenticationManager authenticationManager) {
-        this.jwtTokenizer = jwtTokenizer;
-        this.authorityUtils = authorityUtils;
-        this.logoutListService = logoutListService;
-        this.authenticationManager = authenticationManager;
-    }
-
-//    public JwtVerificationFilter(AuthenticationManager authenticationManager) {
-//        this.authenticationManager = authenticationManager;
-//    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
         try {
-//            AuthenticationManager authenticationManager = builder.getSharedObject(AuthenticationManager.class);
-
-            String bearerToken = request.getHeader("Authorization");
-            String jwt = null;
-
-            if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer")) {
-                Authentication authenticationToken = new JwtAuthenticationToken(bearerToken);
-                Authentication authenticatedToken = authenticationManager.authenticate(authenticationToken);
-                SecurityContextHolder.getContext().setAuthentication(authenticatedToken);
-                jwt = bearerToken.substring(7); // 0~6번째 글자 제외한 문자열
-            }
-
+            String jwt = resolveToken(request);
             String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
 
             if (StringUtils.hasText(jwt) && jwtTokenizer.validateToken(jwt, base64EncodedSecretKey)) {
@@ -131,15 +110,10 @@ public class JwtVerificationFilter extends OncePerRequestFilter {
         SecurityContextHolder.getContext().setAuthentication(authentication);
     }
 
-    private String resolveToken(HttpServletRequest request, HttpSecurity builder) {
-        AuthenticationManager authenticationManager = builder.getSharedObject(AuthenticationManager.class);
-
+    private String resolveToken(HttpServletRequest request) {
         String bearerToken = request.getHeader("Authorization");
 
         if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer")) {
-            Authentication authenticationToken = new JwtAuthenticationToken(bearerToken);
-            Authentication authenticatedToken = authenticationManager.authenticate(authenticationToken);
-            SecurityContextHolder.getContext().setAuthentication(authenticatedToken);
             return bearerToken.substring(7); // 0~6번째 글자 제외한 문자열
         }
